+++ CHANGELOG.md
@@ -71,6 +71,7 @@
 - Replaced REPLACE INTO queries with prepared statements !1124 #757 @alex.simm
 - Renamed "Verschwendungsfasten" to "foodsharing-Städte" @D0nPiano
 - Replace SettingsModel by SettingsGateway !1163 #9 @svenpascal
+- Moved mayHandleReports() to ReportPermissions class with deprecation in Session class !1241 @jofranz
 
 ## Dev/Test/CI stuff